requirements.txt
Django==1.6.5
django-crispy-forms==1.4.0
Pillow==2.5.1
South>=1.0
